Fixed all the merging conflicts
