+++ lib/Settings.dart
@@ -2,6 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:islami/BottomNavBar.dart';
 import 'package:islami/AppConfig.dart';
+import 'package:islami/utility/islami_scaffold.dart';
 import 'package:provider/provider.dart';
 
 class Settings extends StatefulWidget {
@@ -19,105 +20,95 @@ class _SettingsState extends State<Settings> {
 
   @override
   Widget build(BuildContext context) {
+    final isDarkMode = Theme.of(context).brightness == Brightness.dark;
+    isDark = isDarkMode;
     provider = Provider.of<AppConfigProvider>(context);
-    return Scaffold(
-      body: Column(
-        children: [
-          Container(
-            width: MediaQuery.of(context).size.width,
-            height: MediaQuery.of(context).size.height * 0.88,
-            child: Stack(children: [
-              Container(
-                child: Image.asset(
-                  'assets/images/bg3.png',
-                  fit: BoxFit.fitWidth,
-                ),
-                width: MediaQuery.of(context).size.width,
-                height: MediaQuery.of(context).size.height * 0.88,
-              ),
-              Column(
-                children: [
-                  Padding(
-                    padding: const EdgeInsets.fromLTRB(0, 15, 0, 15),
-                    child: InkWell(
-                      child: Row(
-                        children: [
-                          Padding(
-                            padding: const EdgeInsets.fromLTRB(7.0, 0, 8.0, 0),
-                            child: Image.asset('assets/images/globe.png'),
-                          ),
-                          Text(
-                            'Language',
-                            style: TextStyle(
-                                fontSize: 22, fontWeight: FontWeight.bold),
-                            textAlign: TextAlign.start,
-                          ),
-                          Expanded(
-                            child: Padding(
-                              padding: const EdgeInsets.fromLTRB(0, 0, 8, 0),
-                              child: Text(
-                                currentChosenLanguage,
-                                style: TextStyle(
-                                  fontSize: 22,
-                                ),
-                                textAlign: TextAlign.end,
-                              ),
-                            ),
-                          )
-                        ],
+    return IslamiScaffold(
+      bottomNavBarCurrentIndex: BottomNavBar.settingsIndex,
+      child: Container(
+        width: MediaQuery.of(context).size.width,
+        height: MediaQuery.of(context).size.height * 0.88,
+        child: Stack(children: [
+          Column(
+            children: [
+              Padding(
+                padding: const EdgeInsets.fromLTRB(0, 15, 0, 15),
+                child: InkWell(
+                  child: Row(
+                    children: [
+                      Padding(
+                        padding: const EdgeInsets.fromLTRB(7.0, 0, 8.0, 0),
+                        child: Image.asset('assets/images/globe.png'),
                       ),
-                      onTap: onLanguageChange,
-                    ),
-                  ),
-                  Row(children: [
-                    Padding(
-                      padding: const EdgeInsets.fromLTRB(8.0, 0, 8.0, 0),
-                      child: Image.asset(
-                        'assets/images/night_mode.png',
+                      Text(
+                        'Language',
+                        style: TextStyle(
+                            fontSize: 22, fontWeight: FontWeight.bold),
+                        textAlign: TextAlign.start,
                       ),
-                    ),
-                    Text(
-                      'Theme',
-                      style:
-                          TextStyle(fontSize: 22, fontWeight: FontWeight.bold),
-                      textAlign: TextAlign.start,
-                    ),
-                    Expanded(
-                      child: Container(
-                        alignment: Alignment.topRight,
+                      Expanded(
                         child: Padding(
-                          padding: const EdgeInsets.all(8.0),
-                          child: Switch(
-                            value: isDark,
-                            activeColor: Colors.white,
-                            onChanged: (value) {
-                              setState(() {
-                                isDark = value;
-                                if (isDark) {
-                                } else {}
-                              });
-                            },
+                          padding: const EdgeInsets.fromLTRB(0, 0, 8, 0),
+                          child: Text(
+                            currentChosenLanguage,
+                            style: TextStyle(
+                              fontSize: 22,
+                            ),
+                            textAlign: TextAlign.end,
                           ),
                         ),
-                      ),
-                    )
-                  ]),
-                ],
-                crossAxisAlignment: CrossAxisAlignment.stretch,
-                mainAxisAlignment: MainAxisAlignment.center,
+                      )
+                    ],
+                  ),
+                  onTap: onLanguageChange,
+                ),
               ),
-            ]),
+              Row(children: [
+                Padding(
+                  padding: const EdgeInsets.fromLTRB(8.0, 0, 8.0, 0),
+                  child: Image.asset(
+                    'assets/images/night_mode.png',
+                  ),
+                ),
+                Text(
+                  'Theme',
+                  style:
+                  TextStyle(fontSize: 22, fontWeight: FontWeight.bold),
+                  textAlign: TextAlign.start,
+                ),
+                Expanded(
+                  child: Container(
+                    alignment: Alignment.topRight,
+                    child: Padding(
+                      padding: const EdgeInsets.all(8.0),
+                      child: Switch(
+                        value: isDark,
+                        activeColor: Colors.white,
+                        onChanged: (value) {
+                          setState(() {
+                            isDark = value;
+                            if (isDark) {
+                            } else {}
+                          });
+                        },
+                      ),
+                    ),
+                  ),
+                )
+              ]),
+            ],
+            crossAxisAlignment: CrossAxisAlignment.stretch,
+            mainAxisAlignment: MainAxisAlignment.center,
           ),
-          Container(
-            child: BottomNavBar(4),
-          )
-        ],
+        ]),
       ),
     );
+
+
   }
 
   void onLanguageChange() {
-    showModalBottomSheet<void>(
+    showModalBottomSheet(
         context: context,
         backgroundColor: Colors.transparent,
         builder: (context) => Container(
